Actually do far-clipping in hit.
I introduced a mistake when I expressed the `hit_slice` algorithm as a
fold: the C++ code keeps track of the nearest hit so far, and adjusts
the t_range to reflect that. My Rust implementation was always using the
initial t_range, which resulted in spurious hits.

This improves performance *very slightly.*